requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
pyserial>=3.5
websockets>=11.0.3
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.2
httpx>=0.25.0
pytest-asyncio>=0.21.1
